fix: Add NettyExtensionsKtTest

+++ app/src/main/kotlin/com/github/l34130/netty/dbgw/app/handler/codec/mysql/PacketDecoder.kt
@@ -20,7 +20,7 @@ class PacketDecoder : ByteToMessageDecoder() {
         val payloadLength = `in`.readFixedLengthInteger(3)
         val sequenceId = `in`.readFixedLengthInteger(1)
 
-        if (`in`.readableBytes() < payloadLength) {
+        if (`in`.readableBytes().toULong() < payloadLength) {
             // Not enough bytes to read the complete payload.
             `in`.resetReaderIndex()
             return
